also split out spack.mirrors.layout
